fix(registry): classify fixture-gap scenarios untested
