+++ server/src/addie/services/compliance-testing.ts
@@ -530,6 +530,28 @@ function isRunnerApplicabilitySkip(step: {
   }
 }
 
+export function isNonExecutableCoverageGapScenario(scenario: {
+  scenario?: unknown;
+  steps?: Array<{
+    skipped?: boolean;
+    skip_reason?: string;
+    step_id?: unknown;
+    requirement?: unknown;
+  }>;
+}): boolean {
+  const steps = Array.isArray(scenario.steps) ? scenario.steps : [];
+  const scenarioId = typeof scenario.scenario === 'string' ? scenario.scenario : '';
+  return steps.length > 0 && steps.every((step) => {
+    if (!step?.skipped) return false;
+    return step.skip_reason === 'peer_branch_taken' ||
+      step.skip_reason === 'peer_substituted' ||
+      step.skip_reason === 'missing_test_controller' ||
+      step.skip_reason === 'fixture_unavailable' ||
+      (step.skip_reason === 'requirement_unmet' && step.requirement === 'controller') ||
+      isRunnerApplicabilitySkip(step, scenarioId);
+  });
+}
+
 function skipReasonIsCoverageGap(
   reason: string | undefined,
   step?: {
