+++ server/src/routes/registry-api.ts
@@ -35,6 +35,7 @@ import {
 import {
   comply,
   complianceResultToDbInput,
+  isNonExecutableCoverageGapScenario,
   classifyCapabilityResolutionError,
   presentCapabilityResolutionError,
   computeSpecialismStatus,
@@ -8030,35 +8031,6 @@ export function createRegistryApiRouters(config: RegistryApiConfig): { router: R
           steps_total: 0,
         };
         const serializedStoryboardStatus = serializeStoryboardRunStatus(storyboardStatus);
-        const isRunnerApplicabilitySkip = (scenarioId: string, step: {
-          skip_reason?: string;
-          step_id?: unknown;
-          requirement?: unknown;
-        }): boolean => {
-          switch (step.skip_reason) {
-            case 'capability_unsupported':
-              return true;
-            case 'missing_test_kit_contract':
-              return scenarioId === 'idempotency/rate_limit_replay_invariant' &&
-                step.step_id === 'expect_rate_limit_not_replayed';
-            case 'requirement_unmet':
-              return step.requirement === 'webhook_receiver';
-            default:
-              return false;
-          }
-        };
-        const isControllerCoverageGapScenario = (scenario: { scenario?: unknown; steps?: any[] }): boolean => {
-          const steps = Array.isArray(scenario.steps) ? scenario.steps : [];
-          const scenarioId = typeof scenario.scenario === 'string' ? scenario.scenario : '';
-          return steps.length > 0 && steps.every((step) => {
-            if (!step?.skipped) return false;
-            return step.skip_reason === 'peer_branch_taken' ||
-              step.skip_reason === 'peer_substituted' ||
-              step.skip_reason === 'missing_test_controller' ||
-              (step.skip_reason === 'requirement_unmet' && step.requirement === 'controller') ||
-              isRunnerApplicabilitySkip(scenarioId, step);
-          });
-        };
         const uiDiagnostics = (dbInput.step_diagnostics ?? []).map((diagnostic) => ({
           run_id: run.id,
           agent_url: agentUrl,
@@ -8082,7 +8054,9 @@ export function createRegistryApiRouters(config: RegistryApiConfig): { router: R
                   t.scenarios.filter((s) => s.scenario === step.comply_scenario),
                 )
               : [];
-            const executableScenarios = matchingScenarios.filter((s) => !isControllerCoverageGapScenario(s));
+            const executableScenarios = matchingScenarios.filter(
+              (s) => !isNonExecutableCoverageGapScenario(s),
+            );
 
             const passed = executableScenarios.length > 0
               ? executableScenarios.every((s) => s.overall_passed)
